Refactored Brazil Social Security logic to outside of Boleto
