+++ packages/lib/src/components/Boleto/components/BoletoInput/utils.ts
@@ -1,4 +1,4 @@
-import { formatCPFCNPJ } from '../SocialSecurityNumberBrazil/utils';
+import { formatCPFCNPJ } from '../../../internal/SocialSecurityNumberBrazil/utils';
 
 export const boletoFormatters = {
     socialSecurityNumber: ssn => formatCPFCNPJ(ssn)
