--- packages/lib/src/components/Boleto/components/SocialSecurityNumberBrazil/constants.ts
@@ -1,1 +1,0 @@
-export const CPF_LENGTH = 11;
