--- packages/lib/src/components/Boleto/components/SocialSecurityNumberBrazil/validate.ts
@@ -1,3 +1,0 @@
-export default function(ssn) {
-    return /(^\d{3}\.\d{3}\.\d{3}-\d{2}$)|(^\d{2}\.\d{3}\.\d{3}\/\d{4}-\d{2}$)/.test(ssn);
-}
